+++ x/authz/module/abci.go
@@ -8,5 +8,5 @@ import (
 // BeginBlocker is called at the beginning of every block
 func BeginBlocker(ctx sdk.Context, keeper keeper.Keeper) error {
 	// delete all the mature grants
-	return keeper.DequeueAndDeleteExpiredGrants(ctx)
+	return keeper.DequeueAndDeleteExpiredGrants(ctx, 200)
 }